pyproject.toml
[project]
name = "eoepca-security"
version = "0.2.0"
description = "Utilities for writing Python clients for use with the EOEPCA IAM"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=44.0.1",
    "fastapi[standard]>=0.115.8",
    "pyjwt>=2.10.1",
    "requests>=2.32.3",
    "requests-cache>=1.2.1",
    "types-requests>=2.32.0.20241016",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "pytest>=8.3.4",
    "pytest-mypy>=0.10.3",
    "pytest-ruff>=0.4.1",
    "ruff>=0.9.8",
]

[tool.pytest.ini_options]
addopts = "-x --ruff --ruff-format"

[tool.mypy]
strict = true
